refactor the Toolbat
