+++ src/components/story-builder/menus/index.tsx
@@ -1,0 +1,5 @@
+export { FileMenu } from './FileMenu';
+export { ExportMenu } from './ExportMenu';
+export { SceneMenu } from './SceneMenu';
+export { AssetsMenu } from './AssetsMenu';
+export { ViewSelector } from './ViewSelector';
